Remove visibility label from row metric
This was causing an issue when a user is a very small subset of the
entire table, and the query would be returning a large portion of the
rows.  The query would scan say 30 rows of the 50 that the user can see,
but the threshold limit would cause it to skip the index.  However, the
actual table can contain millions of rows, invoking a full table scan.
The correct method is to leave the number of rows in the table as such,
so the calculation will be accurate and the index will be used.