Add PrestoBatchWriter
This commit simplifies the ingestion API for both the
connector as well as external applications that use this
code as a library.  Applications can use this class
to add new rows, delete rows, and update columns of existing rows,
without having to deal with the index and metric entries as well.