Use LoadingCache#get if there is only one exact range
Exact ranges of a predicate are pulled from the LoadingCache via getAll,
however we've seen an issue where concurrent threads are querying
a single exact value and multiple threads reach out to Accumulo for
the value.  When using LoadingCache#get, one thread would load the
value while the other threads block waiting for the value to be
loaded -- getAll does not share this behavior.  This commit uses
'get' over 'getAll' when there is only one exact range, removing
the redundant fetches from Accumulo.