Add SortedEntryIterator wrapper
This wrapper fixes an issue where some results returned by the
AccumuloRecordCursor were incomplete.  Due to the use of a BatchScanner,
the entries returned could span multiple rows, causing incorrect
results.  By wrapping the BatchScanner's Iterator in this new iterator,
we re-order the stream of results to fix the functionality of the
RecordCursor.