Update PrestoBatchWriter#addMutation to support delete entries
This addresses an issue where Indexer#index rejects a mutation that contains
any 'delete' entries.  Since the PrestoBatchWriter uses this,
end users can call addMutation that is a mix of put and delete
entries and it would be rejected by the Indexer.  Instead, we'll
split the Mutation into two separate entries and call a new
Indexer#delete that accepts a Mutation and only deletes the
corresponding index/metric entries.  See the Javadoc comments for usage.